requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0
django-filter==23.5
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
drf-nested-routers==0.93.5
idna==3.6
iniconfig==2.0.0
kombu==5.3.7
Markdown==3.5.2
oauthlib==3.2.2
packaging==24.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.4
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
